webportal/requirements.txt
Flask==3.1.0
humanize==4.12.3
jsonschema==4.23.0
Markdown==3.8
pymongo==4.12.1
Requests==2.32.3
yarl==1.20.0
gunicorn==21.2.0
pyyaml==6.0.2